+++ package.json
@@ -11,7 +11,9 @@
     "@babylonjs/core": "^7.51.0",
     "@babylonjs/gui": "^7.51.0",
     "@babylonjs/loaders": "^7.51.0",
-    "vue": "^3.2.13"
+    "vue": "^3.2.13",
+    "vuetify": "^3.5.9",
+    "@mdi/font": "^7.4.47"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.4.0",
@@ -22,6 +24,6 @@
     "@vue/eslint-config-typescript": "^9.1.0",
     "eslint": "^7.32.0",
     "eslint-plugin-vue": "^8.0.3",
-    "typescript": "~4.5.5"
+    "typescript": "~4.7.4"
   }
 }
