+++ src/babylon/components/circularRooms.ts
@@ -110,4 +110,4 @@ export default class CircularSeating {
 
         return outObjects;
     }
-}
+}
