+++ src/babylon/components/twoCube.ts
@@ -31,4 +31,4 @@ export default class TwoCube {
         return seats;
 
     }
-}
+}
