+++ src/main.ts
@@ -1,4 +1,11 @@
 import { createApp } from 'vue'
-import App from './App.vue'
+import App from './vue/App.vue'
+import { createVuetify } from 'vuetify'
+import 'vuetify/styles'
+import '@mdi/font/css/materialdesignicons.css'
 
-createApp(App).mount('#app')
+const vuetify = createVuetify()
+
+createApp(App)
+  .use(vuetify)
+  .mount('#app')
