+++ src/shims-vuetify.d.ts
@@ -1,0 +1,6 @@
+declare module 'vuetify' {
+  import { Framework } from 'vuetify'
+  const Vuetify: Framework
+  export default Vuetify
+  export { createVuetify } from 'vuetify'
+}
