+++ src/vue/App.vue
@@ -1,15 +1,19 @@
 <template>
-  <HelloWorld msg="Welcome to Your Vue.js + TypeScript App"/>
+  <v-app>
+    <v-main>
+      <ProceduralScene />
+    </v-main>
+  </v-app>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import HelloWorld from './components/HelloWorld.vue';
+import ProceduralScene from './components/ProceduralScene.vue';
 
 export default defineComponent({
   name: 'App',
   components: {
-    HelloWorld
+    ProceduralScene
   }
 });
 </script>
@@ -23,4 +27,4 @@ export default defineComponent({
   color: #2c3e50;
   margin-top: 60px;
 }
-</style>
+</style>
